Move NIP-04 detection into nostr/protocol
