+++ web/src/lib/nostr/protocol/nip04.ts
@@ -1,0 +1,1 @@
+export const isLegacyEncryption = (content: string): boolean => content.includes('?iv=');
